Use unsigned int where possible

+++ src/components/model/mesh.cpp
@@ -4,7 +4,7 @@
 #include "log.hpp"
 
 Mesh::Mesh(const std::vector<Vertex> &vertices, const std::vector<unsigned int> &indices, const std::vector<Material> &materials)
-    : _vertices(vertices), _indices(indices), _materials(materials), _indexCount((uint32_t)indices.size())
+    : _vertices(vertices), _indices(indices), _materials(materials), _indexCount(indices.size())
 {
     InitializeVertexBuffer();
     InitializeIndexBuffer();
@@ -23,7 +23,7 @@ void Mesh::Render() const
 void Mesh::InitializeVertexBuffer()
 {
     D3D11_BUFFER_DESC bufferDesc = {};
-    bufferDesc.ByteWidth = (uint32_t)_vertices.size() * VERTEX_STRIDE;
+    bufferDesc.ByteWidth = _vertices.size() * VERTEX_STRIDE;
     bufferDesc.StructureByteStride = VERTEX_STRIDE;
     bufferDesc.Usage = D3D11_USAGE_DEFAULT;
     bufferDesc.BindFlags = D3D11_BIND_VERTEX_BUFFER;
@@ -38,7 +38,7 @@ void Mesh::InitializeVertexBuffer()
 void Mesh::InitializeIndexBuffer()
 {
     D3D11_BUFFER_DESC bufferDesc = {};
-    bufferDesc.ByteWidth = (uint32_t)_indices.size() * sizeof(unsigned int);
+    bufferDesc.ByteWidth = _indices.size() * sizeof(unsigned int);
     bufferDesc.Usage = D3D11_USAGE_DEFAULT;
     bufferDesc.BindFlags = D3D11_BIND_INDEX_BUFFER;
